+++ app/src/test/java/woowacourse/movie/domain/ScreeningTimeTest.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie
+package woowacourse.movie.domain
 
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Test
